修改week3至稳定版
修改week3至稳定版,推荐使用

+++ Project/WEEK3/src/init.h
@@ -29,5 +29,5 @@
    * Êä³ö : ÎÞ
    ***********************************************************************/
   extern void beep_Init(void);
-  extern int abs(int a);
+  extern float abs(float a);
 #endif
